skaffold-STABLE/webapp/requirements.txt
Flask==1.0.2
emoji
flask_sqlalchemy
SQLAlchemy
psycopg2-binary
pymemcache.client
#time
essential_generators
kafka-python